Introduce SimpleForm component
