+++ src/mui/detail/Create.js
@@ -1,11 +1,9 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { Card, CardTitle, CardActions } from 'material-ui/Card';
+import { Card, CardTitle } from 'material-ui/Card';
 import inflection from 'inflection';
 import Title from '../layout/Title';
-import ListButton from '../button/ListButton';
 import { crudCreate as crudCreateAction } from '../../actions/dataActions';
-import RecordForm from './RecordForm'; // eslint-disable-line import/no-named-as-default
 import DefaultActions from './CreateActions';
 import getDefaultValues from './getDefaultValues';
 
@@ -18,7 +16,7 @@ class Create extends Component {
     handleSubmit = (record) => this.props.crudCreate(this.props.resource, record, this.getBasePath());
 
     render() {
-        const { actions = <DefaultActions/>, children, defaultValue = {}, isLoading, resource, title, validation } = this.props;
+        const { actions = <DefaultActions />, children, defaultValue = {}, isLoading, resource, title } = this.props;
         const basePath = this.getBasePath();
         return (
             <Card style={{ margin: '2em', opacity: isLoading ? 0.8 : 1 }}>
@@ -27,16 +25,13 @@ class Create extends Component {
                     resource,
                 })}
                 <CardTitle title={<Title title={title} defaultTitle={`Create ${inflection.humanize(inflection.singularize(resource))}`} />} />
-                <RecordForm
-                    onSubmit={this.handleSubmit}
-                    resource={resource}
-                    basePath={basePath}
-                    validation={validation}
-                    record={{}}
-                    initialValues={getDefaultValues(children)({}, defaultValue)}
-                >
-                    {children}
-                </RecordForm>
+                {React.cloneElement(children, {
+                    onSubmit: this.handleSubmit,
+                    resource,
+                    basePath,
+                    record: {},
+                    initialValues: getDefaultValues(children)({}, defaultValue),
+                })}
             </Card>
         );
     }
@@ -44,7 +39,7 @@ class Create extends Component {
 
 Create.propTypes = {
     actions: PropTypes.element,
-    children: PropTypes.node,
+    children: PropTypes.element,
     crudCreate: PropTypes.func.isRequired,
     defaultValue: PropTypes.object,
     isLoading: PropTypes.bool.isRequired,
@@ -52,7 +47,6 @@ Create.propTypes = {
     params: PropTypes.object.isRequired,
     resource: PropTypes.string.isRequired,
     title: PropTypes.any,
-    validation: PropTypes.func,
 };
 
 Create.defaultProps = {
