+++ src/mui/detail/index.js
@@ -1,4 +1,4 @@
 export Create from './Create';
 export Edit from './Edit';
-export RecordForm from './RecordForm';
+export SimpleForm from './SimpleForm';
 export Show from './Show';
